Refactor out DefenceTpe and add response package
