+++ ccd-adapter/src/test/java/uk/gov/hmcts/cmc/ccd/mapper/response/ResponseMapperTest.java
@@ -8,8 +8,9 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import uk.gov.hmcts.cmc.ccd.config.CCDAdapterConfig;
 import uk.gov.hmcts.cmc.ccd.domain.response.CCDResponse;
-import uk.gov.hmcts.cmc.domain.models.FullDefenceResponse;
-import uk.gov.hmcts.cmc.domain.models.Response;
+import uk.gov.hmcts.cmc.domain.models.response.DefenceType;
+import uk.gov.hmcts.cmc.domain.models.response.FullDefenceResponse;
+import uk.gov.hmcts.cmc.domain.models.response.Response;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleResponse;
 
 import static uk.gov.hmcts.cmc.ccd.assertion.Assertions.assertThat;
@@ -38,7 +39,7 @@ public class ResponseMapperTest {
     public void shouldMapAlreadyPaidFullDefenceResponseToCCD() {
         //given
         FullDefenceResponse fullDefenceResponse = (FullDefenceResponse) SampleResponse.FullDefence.builder()
-            .withDefenceType(FullDefenceResponse.DefenceType.ALREADY_PAID)
+            .withDefenceType(DefenceType.ALREADY_PAID)
             .withMediation(Response.FreeMediationOption.NO)
             .withMoreTimeNeededOption(Response.MoreTimeNeededOption.NO)
             .build();
