+++ ccd-claim-migration/src/main/java/uk/gov/hmcts/cmc/ccd/migration/mappers/ClaimMapper.java
@@ -6,8 +6,8 @@ import uk.gov.hmcts.cmc.ccd.migration.config.JacksonConfiguration;
 import uk.gov.hmcts.cmc.domain.models.Claim;
 import uk.gov.hmcts.cmc.domain.models.ClaimData;
 import uk.gov.hmcts.cmc.domain.models.CountyCourtJudgment;
-import uk.gov.hmcts.cmc.domain.models.Response;
 import uk.gov.hmcts.cmc.domain.models.offers.Settlement;
+import uk.gov.hmcts.cmc.domain.models.response.Response;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
