+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/response/DefenceType.java
@@ -1,0 +1,16 @@
+package uk.gov.hmcts.cmc.domain.models.response;
+
+public enum DefenceType {
+    DISPUTE("I dispute all the claim"),
+    ALREADY_PAID("I have paid what I believe I owe");
+
+    String description;
+
+    DefenceType(String description) {
+        this.description = description;
+    }
+
+    public String getDescription() {
+        return this.description;
+    }
+}
