+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/response/DefendantTimeline.java
@@ -1,7 +1,8 @@
-package uk.gov.hmcts.cmc.domain.models;
+package uk.gov.hmcts.cmc.domain.models.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
+import uk.gov.hmcts.cmc.domain.models.TimelineEvent;
 
 import java.util.Collections;
 import java.util.List;
