+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/response/FullDefenceResponse.java
@@ -1,6 +1,7 @@
-package uk.gov.hmcts.cmc.domain.models;
+package uk.gov.hmcts.cmc.domain.models.response;
 
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
+import uk.gov.hmcts.cmc.domain.models.PaymentDeclaration;
 import uk.gov.hmcts.cmc.domain.models.evidence.DefendantEvidence;
 import uk.gov.hmcts.cmc.domain.models.legalrep.StatementOfTruth;
 import uk.gov.hmcts.cmc.domain.models.party.Party;
@@ -14,21 +15,6 @@ import javax.validation.constraints.Size;
 import static uk.gov.hmcts.cmc.domain.utils.ToStringStyle.ourStyle;
 
 public class FullDefenceResponse extends Response {
-
-    public enum DefenceType {
-        DISPUTE("I dispute all the claim"),
-        ALREADY_PAID("I have paid what I believe I owe");
-
-        String description;
-
-        DefenceType(String description) {
-            this.description = description;
-        }
-
-        public String getDescription() {
-            return this.description;
-        }
-    }
 
     @NotNull
     private final DefenceType defenceType;
