+++ src/aat/java/uk/gov/hmcts/cmc/claimstore/tests/helpers/CommonOperations.java
@@ -73,7 +73,7 @@ public class CommonOperations {
     }
 
     public Response submitResponse(
-        uk.gov.hmcts.cmc.domain.models.Response response,
+        uk.gov.hmcts.cmc.domain.models.response.Response response,
         String claimExternalId,
         User defendant
     ) {
