+++ src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/controllers/IsDefendantLinkedTest.java
@@ -8,7 +8,7 @@ import uk.gov.hmcts.cmc.claimstore.BaseIntegrationTest;
 import uk.gov.hmcts.cmc.claimstore.idam.models.User;
 import uk.gov.hmcts.cmc.claimstore.services.notifications.fixtures.SampleUserDetails;
 import uk.gov.hmcts.cmc.domain.models.Claim;
-import uk.gov.hmcts.cmc.domain.models.DefendantLinkStatus;
+import uk.gov.hmcts.cmc.domain.models.response.DefendantLinkStatus;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaimData;
 
 import static org.assertj.core.api.Assertions.assertThat;
