+++ src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/services/staff/DefendantResponseStaffNotificationServiceTest.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import uk.gov.hmcts.cmc.claimstore.MockSpringTest;
 import uk.gov.hmcts.cmc.claimstore.config.properties.emails.StaffEmailProperties;
 import uk.gov.hmcts.cmc.domain.models.Claim;
-import uk.gov.hmcts.cmc.domain.models.FullDefenceResponse;
+import uk.gov.hmcts.cmc.domain.models.response.DefenceType;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaim;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleResponse;
 import uk.gov.hmcts.cmc.email.EmailAttachment;
@@ -49,7 +49,7 @@ public class DefendantResponseStaffNotificationServiceTest extends MockSpringTes
             .withResponse(
                 SampleResponse.FullDefence
                     .builder()
-                    .withDefenceType(FullDefenceResponse.DefenceType.ALREADY_PAID)
+                    .withDefenceType(DefenceType.ALREADY_PAID)
                     .withMediation(null)
                     .build()
             )
